backendAuth/requirements.txt
fastapi
uvicorn
psycopg2-binary
pyjwt
passlib
bcrypt==3.2.2
python-multipart